Clean up clean script by reusing code
The only user-visible change should be that this causes the list of
targets to be the same between the build scripts and the clean script.

Internally the `_common.sh` script is reused to avoid duplicating
logic. To achieve this, that script now defines only environment
variables and functions but does not do anything actively. All the
build initialization has been moved to a function `build_setup` to which
calls are added where needed.